Declare dependencies with version ranges rather than single versions
…and lock exact versions in a simple, hopefully obvious
requirements-locked.txt file.  Notably, the exact versions are locked
with `pip freeze` so that transitive deps are included as well.  This
helps install in more diverse environments and should require less
manual co-ordination between sacra, fauna, and augur.  It also more
precisely pins all dep versions when using the lock file.

I generated the requirements-locked.txt using `pip freeze --local` in a
clean virtualenv after installing the updated requirements.txt.  The
version ranges were determined after manually checking the versioning
policies or practices of our deps.

The general idea here is based on "A Better Pip Workflow".¹  While
Pipenv now exists for this problem space, it's an additional layer of
tooling that we'd like to see if we can avoid for now.

¹ https://www.kennethreitz.org/essays/a-better-pip-workflow