Remove treetime_augur submodule
While a custom version was necessary at one point, this vendored code
has been pointing at the TreeTime mainline for over a year, since the
commit "move treetime back to master" (32230e4).

Remove it and the complications of submodules in favor of requiring
TreeTime as a (mostly) normal dep in requirements.txt.

Note that this change depends on PR #44 in TreeTime¹ which contains the
commit "Declare deps in setup.py so this is an installable library".

¹ https://github.com/neherlab/treetime/pull/44