Read default data files using pkg_resources
Using pkg_resources is the recommended practice:

   https://setuptools.readthedocs.io/en/latest/setuptools.html#accessing-data-files-at-runtime

This change moves the locating of the data files into the
read_lat_longs() and read_colors() functions themselves to make the
caller's code simpler.  This is in part because it's best practice to
request package resources as a pre-read data string or an open stream,
rather than an actual filename.  See also:

   http://peak.telecommunity.com/DevCenter/PythonEggs#accessing-package-resources