Check all shell command invocations for failure
Replaces os.system() calls with a small utility function,
run_shell_command(), which prints an error message if the command exits
with failure and returns or raises an exception appropriately.  This is
a thin wrapper around subprocess.check_call(), which does most of the
heavy-lifting.  subprocess is Python's modern collection of functions
for running external commands.

It is better to catch command failure early, immediately after it
happens, than to proceed on the assumption that everything worked and
discover failure later via proxies like missing or empty files.  Doing
so makes what failed and why more clear.

This only changes modular augur.  In the changed callsites, the newly
introduced exceptions are only caught if existing exceptions were being
caught and handled.