Added titer model predictor that censors future data
The new predictor selects all titer measurements between the date of the tree's
root node and the current timepoint, reruns the requested titer model with this
new subset of date, and reannotates the tree with the corresponding titer drops.

This implementation repurposes the subset_to_date method from TiterCollection to
take advantage of the existing filter_strains method.