Optionally export amino acid translations to auspice tree JSONs
By exporting these translations to the auspice tree JSONs, the titer
substitution model no longer needs to reconstruct pairwise amino acid
differences between nodes in a tree. The code to perform this reconstruction is
now available as a gist instead:

https://gist.github.com/huddlej/4e43f0b6b9f8e92cb5155b6372b11182

Although this approach modifies the "old augur" process interface, it should be
consistent with the modular augur interface which already allows for exporting
amino acid translations.