Add release tooling
The devel/release script sanity checks state of the repo and change log,
bumps the version, makes a commit and tags it, merges into the "release"
branch, and ends with instructions on how to push.  It is based on the
very similar script I wrote for nextstrain-cli, with the "release"
branch behaviour adopted from auspice.

Resolves #202 and #146.