Export amino acid translations in JSON format
Follow the convention from augur process of creating a per-node "translations"
attribute that contains the full amino acid sequence per feature. The resulting
JSON can be merged with other node-specific annotations in the auspice
export. These amino acid translations are required to calculate scores like
epitope mutations for flu.

Since translations are stored by gene in a dictionary, the coordinates of each
gene in a segment are required to restore the complete amino acid sequence. This
is primarily an issue for HA builds with flu right now.