Move the top-level command's logic into augur.run()
This function is called from main() in augur.__main__, which is the
setuptools entrypoint for the installed "augur" command.  Entrypoints
are the recommended way to install commands for several reasons, the
primary one being that they take care of platform-specific requirements.
The __main__ module is also the way that commands are invoked via
"python -m", so this change doubles as allowing that method to work.

bin/augur is now a development-only script — it isn't installed — which
will always load the augur code in the repo instead of whatever version
of augur may be installed.  This means it'll work for testing changes
without needing to (re-)install augur (editable with -e or otherwise).