Revamp README now that we're releasing versioned packages
The aim was to make the README suitable for publishing in the
nextstrain-augur package and display on PyPi.  Part of that was making
it more cohesive and focus more on modular Augur instead of older Augur.

I've adopted Trevor's preferred single-sentence per line style for the
sections I was editing (most of them).