setup.py: Reconcile conflicting license declaration in metadata vs. README/LICENSE.txt
The original license was AGPLv3, but the package metadata switched to
MIT in 818b39a.  Correct the metadata to agree with the README.md and
LICENSE.txt files.

Note that the "license" field is apparently for when there is not an
appropriate classifier string for the license in use.¹  This was new to
me, but since the doc is clear, I've removed it.

¹ https://docs.python.org/3/distutils/setupscript.html#additional-meta-data