Optionally auto-scale the number of threads used for alignment and tree building
Passing "--nthreads auto" to the "align" or "tree" commands will use all
the available CPU cores.  The default is reduced from 2 to 1, which is
the only safe assumption.

Auto-scaling is not the default because it's not necessarily reasonable
to assume that all cores should be used without more information about
the system's workload.  For example, someone running two simple `augur
tree` commands may not reasonably expect them to conflict with each
other, but they would with auto-scaling enabled by default.  Parallelism
requires co-operation and thus an explicit opt-in.