Do not negate non-epitope distances from ancestral nodes
In the original Luksza and Lassig model, non-epitope mutations are expected
to contribute negatively to a clade's overall fitness, so the number of
mutations between a strain and its closest ancestor is multiplied by -1. Here,
I'm choosing to keep the non-epitope distances positive for better
interpretability of the standardized values and allowing the beta parameter
assigned to this predictor to be negative if it needs to be.