Use censored frequencies for observed future clade frequencies
The reasoning here is that even "in the future", we still have uncertainty about
what true clade frequencies are and that uncertainty is not resolved until some
time past the delta projection time. Using uncensored frequencies makes the
problem the model tries to solve that much more difficult and also different
from the problem it is designed to solve.