Fixed a bug with observed final clade frequencies
I originally used the wrong index offset into the timepoint array to get the
"future" frequencies associated with a given timepoint. This caused observed
future clade frequencies to be sourced from one timepoint in the future (6
months) instead of delta time in the future (1 year).

This commit fixes this bug and also replaces the timepoint calculations
themselves with a datetime-aware range function from pandas that obviates the
accumulation of floating rounding errors associated with the original
numpy-based approach. This latter change was necessary to enable lookup of
future frequencies with a simple time + delta_time calculation that was not
previously possible due to accumulated rounding errors in timepoints.