Travis CI: Fix silently failing test and release scripts
Restores the sub-shells that the tests use to isolate directory changes.

The commit "Include Travis environment YAML and update Travis build"
(1cfa782) simplified the Travis config by introducing a conda
environment file and removing testing under Python 2.7.  A subtle
removal of parentheses caused the two "script" lines to be run in the
top-level shell instead of sub-shells.  This made the effect of changing
directories global, which meant the first `cd` worked but the subsequent
one did not and the invocation of ./devel/travis-rebuild-docker-image
similarly failed.