Add KDE estimator for multiple sequence alignments
Adds a new AlignmentKdeFrequencies subclass of KdeFrequencies that accepts a
multiple sequence alignment and a list of observation dates for each sample in
the alignment and returns estimated frequencies at each position and
base/residue in the given alignment for all variant sites.

This commit includes some additional clean up to make the `calculate_pivots`
class method more generic and less aware of trees or other external data
structures.

Still unimplemented is an interface to this new class through the modular augur
command line and several subtle features of the diffusion frequency estimator
like the ability to ignore gap characters, etc. Also missing is the ability to
weight tips in alignment-based estimates.