Enable annotation of nodes in a tree from the substitution model
Adds an optional `--tree` argument to the titer substitution model's subcommand
and corresponding logic to annotate per-branch and cumulative antigenic advance
to each branch in the given tree based on the mutations between a given node and
its parent.

This commit additionally restructures the titer model subparsers to follow the
general augur binary approach of defining a `__command__` per parser that refers
to a class with a `run` method that receives the parser arguments. This
restructuring allows users to run `augur titers` without any arguments and get a
help message instead of an exception (thanks to the helpful
`add_default_command` function).