Update references for H1N1pdm to A/California/07/2009
Reference sequences are copied from the seasonal flu repository [1] and updated
here to replace the previous H1N2 sequences that produced problematic multiple
sequence alignments for NA builds. Since TreeTime is used now to root the tree,
the sequences we use as "outgroups" are really references for the alignments.

Additionally, this commit fixes a bug that caused NA builds for H1N1pdm to break
because there are no epitope masks defined for that lineage. The bug dates back
to July 30, 2018 when I added support of NA epitope masks to H3N2 builds [2].

[1] https://github.com/nextstrain/seasonal-flu/tree/master/config/reference_h1n1pdm*
[2] https://github.com/nextstrain/augur/commit/27d8e15