allow negative conditions in augur filter exclusions
Previously it was possible to exclude sequences in "augur filter" by matching a metadata property via "--exclude-where prop=value". This commit allows a negative match to be used here, i.e. "--exclude-where prop!=value".
Note that this is different to "--include-where prop=value", and the help message for this flag has been improved.