add tree_name property
This is needed for us to display tangle-trees, else there is no way of knowing what portion of the filename is the tree and thus what other datasets are valid as the second tree. Note that previously this was "guessed" by the server.

Add a corresponding command line option to `augur export` to set this