write_json: Remove all indentation by default if AUGUR_MINIFY_JSON is set
This is comparable to export's --minify-json flag, but applies to all
commands and is easily toggled pipeline-wide for dev vs. production.

For parity with John's choice with --minify-json, spaces are left
between and within key/value pairs.  A future enhancement might strip
those too.

It is possible we should also add --minify-json flags to all commands
for consistency with export, but it seems like export is the most useful
command to have the flag.  The files export produces are generally
expected to be longer-lived than the files produced by other commands at
intermediate steps.

Another future enhancement might be to automatically gzip JSON output
(and gunzip JSON input, which would require some more refactoring).