Populate API docs with sphinx-apidoc
sphinx-api doc searches requested paths to modules and creates an autodoc entry
for each module it finds. In this case, we autogenerated an `augur.rst` file and
reference that file's table of contents from a static API.rst file. The augur
API file was generated with the command `sphinx-apidoc -T -f -o docs augur` from
the top-level augur directory.