Add handling for insufficient data.
This adds the --allow-empty-model flag for tree and substitution models, which will handle the InsufficientDataException. Using the flag in a case that raises the exception will result in empty model outputs being written. In cases that raise the exception when the flag is not in use, the program will now exit.