Docs: Fix automatic author list generation for Read The Docs
Read The Docs runs the build process without an attached terminal
(reasonably so), which changes git shortlog's default behaviour when no
revisions are specified.  Instead of defaulting to HEAD, it reads commit
messages from stdin.  No input was redirected to stdin, so git shortlog
read nothing and exited without error, thus silently failing to do our
intent.  This behaviour is documented by git shortlog's manpage, but I
wasn't aware of it.

The fix is to explicitly ask for a shortlog starting from the currently
checked out commit (HEAD), which works as expected regardless of stdin.