Add support for FASTA output to ancestral subcommand
Adds a new flag `--output-sequences` and logic to support saving ancestral sequences
and leaves from the given tree to a FASTA file. Also adds a redundant, more
specific flag `--output-node-data` that will replace the current `--output` flag in
the next major version release of augur. For now, we issue a deprecation warning
when the `--output` flag is used.

Note that FASTA output is only allowed for FASTA inputs and not for VCFs. We
don't allow FASTA output for VCFs anywhere else and, if we did here, the output
files would be very large.

Closes #292.