Fix augur filter arguments, making output required, exclude optional
Requires a value for the `--output` argument instead of allowing users to omit
the value and get a value error exception. Additionally, fixes a bug in the
handling of the excluded strains file when no argument is provided and the
filter logic still tries to test if a None value is a file or not.

Specifically, we now check once whether the exclude argument is given and use
Python's built-in exception to handle the case when the requested file does not
exist. We also report a missing file by printing a helpful message to standard
error and exiting with a non-zero error code instead of raising an exception.

Closes #261.