Refactor logic to read trees from multiple formats into a function
Adds a new function `read_tree` to the `utils` module that tries to safely
handle reading trees in multiple input formats. This kind of try/except approach
to loading trees is necessary because Bio.Phylo.read requires a specified tree
format, but we do not know the format of a given tree by the filename extension
alone.

A side effect of this commit is that missing input files are also now properly
caught and reported to the user on the command line instead of raising a Python
exception.

Closes #302.