Assign authors to nodes & remove author_info metadata
We previously had a mapping of a "citekey" (or similar) attached to a node and a lookup-disctioary of publication info ("author_info"). This was done mainly to save space in the JSON, which is now less of a concern since we almost always use compressed files. By shifting the authorship information to each node we vastly simplify both the JSON and the auspice code needed to visualise it.