improve geographic trait parsing in export v2
Previously `augur export v2` was only looking for the "geo" key in the auspice config (deprecated in v2) - it now also looks for the "geographic_info" key. Additionally, the command line arguments now override any provided in the config (this may want to be tweaked).
Tested using the zika test build.