Run shell commands with more robust error checking
Shell commands are now run under Bash's errexit, nounset, and pipefail
options.  These provide a "strict mode" for Bash that help errors be
caught more reliably and earlier in execution.

The default pipeline error handling tripped up augur filter/mask on VCF
input when vcftools was missing: the command appeared to succeed because
vcftools was invoked as the first part of a multi-command pipeline.

Since we're now explicitly using features of Bash and not defaulting to
/bin/sh, we go the extra mile to report a nicer error if the
near-ubiquitous /bin/bash isn't actually available.

In the future, we'll bypass the shell entirely by switching from shell
command strings to executing programs directly.  For now, that's a
larger change which will have to wait.

Resolves #329.