move to using "geo_resolutions" and ordering them
This changes v2 JSONs, and the auspice config files which `augur export v2` consumes, to store geographic resolutions under the "geo_resolutions" property which is more descriptive. The data structure has changed such that the exported resolutions are ordered (previously we relied on dict ordering which always worked but is not ideal).