Fix missing annotation of "parent" attribute for the root node
Fixes a bug in the annotation of parents to nodes in a tree where the root node
was missing an annotation. This missing annotation broke any code that attempted
to interrogate the parent attribute by looping through all nodes in the tree
with the `find_clades` method. This commit fixes the bug and adds doctests to
verify the code is working.