Parameterize size of clades to estimate by diffusion frequency likelihood
Exposes a previously hardcoded parameter for the diffusion tree frequency
estimator that determines how many tips a clade must have to be considered for
frequency estimation by the diffusion likelihood calculation. Clades that are
smaller than the specified threshold inherit their parent's frequencies
proportionally to their contribution to the parent clade.

This commit exposes the `min_clades` parameter through a keyword argument to the
diffusion frequencies class and also through the `augur frequencies` command
with the new `--minimal-clade-size-to-estimate` argument. This commit modifies
the description of the existing `--minimal-clade-size` argument to clarify its
use for filtering output in contrast to the new argument's use for controlling
frequency estimation itself. The previously hardcoded default value of 10 is now
the default for both the frequencies class and the command line argument.