export v2: Swap "version": "2.0" for "v2"
The exported JSON had set "version" to "2.0". We have decided to explicitly pin the JSON "version" to the version of auspice that reads this JSON. I thought that having a JSON of "2.0" could cause confusion if it's read by auspice version "2.2.0". 

By versioning JSON as "v2" it avoids this confusion.