move docs sections into directories. Update CLI pages.
The content behind each "section" of the docs (e.g. usage, tutorials) is now contained within a directory of the same name. This makes the structure of the `./docs/` directory much easier to understand & makes it more obvious that `index.rst` is the entrypoint for the docs.

The CLI documentation has been updated to include
(a) a few missing pages -- it now represents all augur subcommands
(b) the previously disjoint autogenerated documentation and hand written documentation for a subcommand have been combined into the same file (and therefore link in the docs).

There have been no URL changes with this commit (compared with the currently released docs).