[docs] edits to new export command
I've made somewhat extensive changes to the documentation around `augur export`, but the actual content is mostly preserved.

I've tried to minimise usage of `in-line code` and **bold** text within paragraphs as I don't like how they are rendered in read-the-docs. Similarly, i've changed subsections to often be a stand-alone line of bold text as the font sizes are too similar between h-headings.

I've also tried to minimise auspice implementation details and focus on augur-only content.