[docs] tidy up "using augur" structure
Moves the introduction text to the main "Using augur" page to improve the first impression.

Uses manual TOCs in the pages with examples etc as the autogenerated TOCs weren't aesthetically pleasing -- if you used "heading 1" headers then they appeared in the sidebar (not desired) and if you used "heading 2" then they appeared at the same level as "Named arguments".