Lowercase subsection headings for literal keys in config JSON
To avoid reader confusion, this commit lowercases the subsection headings for
keys like "geo_resolutions" that were previously capitalized. This commit does
not additionally call out these literal values as inline code because doing so
reduces the readability of the documentation. This latter point might be worth
discussing in the future though.