enforce augur (major) version matches, if present
`read_node_data` will now die with a helpful error message if a provided JSON specifies an augur version which has a different major-version-number than the currently running augur. If a augur version is not specified no checks are run. This allows non-augur-produced JSONs to be used without requiring a version number, but also allows augur-v5 JSONs to be used in augur-v6 which needs to be checked for separately.
This commit adds the "packaging" library ("Core utilities for Python packages") to augur