move to validating the annotations block of a JSON
Since we don't know the node-data JSON "type" via the filename, and there is no strong mapping anyway, we cannot define validators for individual JSONs. This leaves us needing either (a) a generic node-data JSON schema or (b) a schema for certain parts of node-data JSONs. This commit implements the latter, as it's far simpler. Should be revisited in the future.