remove unused "tree name" argument from export v2
This argument is not used in `augur export v2`. Using it will now result in a fatal error. While it could be deprecated, and a warning printed, there are multiple arguments from v5's `augur export` (and v6's `augur export v1`) which are not used by `augur export v6` and similarly result in fatal errors. Documentation has been updated to reflect this.