Validate config JSON in `export v2`
Previously this was desired but not implemented due to the backward compatibility of `export v2` in handling v1 config JSONs. This modifies the schema to allow "deprecated" keys & thus allow us to validate the schema programmatically.

Additionally, this modifies the schema to allow both "geo_resolutions" structures which were supported in the `export v2` code. Updates the zika example to use a combination of the two so that the tests cover this.