Fix weighted frequencies when weight keys are unrepresented
Fixes a bug in weighted KDE estimation of tree frequencies when one or more
weight attributes is not represented by any tips on the tree. For example, when
a tree is missing any samples from Africa but the frequencies are being weighted
by region, the total of the estimated frequencies will not sum to 1.

This commit adds unit tests to confirm the original issue and adds logic to KDE
frequency estimation to correct the issue.

Closes #268.