[align] remove temporary files
If a reference sequence file is provided, we create a temporary file containing the input sequences + the reference. This commit removes this file upon completion. This will become important as we extend the functionality to add sequences to an additional alignment to prevent temporary files in the "data" directory in a typical nextstrain directory layout.