[align] allow sequences to be added to existing alignment
Having a hand-curated alignment which is undesirable to change is common for certain pathogens. This commit allows `augur align` to add new data to an existing alignment. 

The trimming of gaps w.r.t. reference is not done if there are no gaps introduced in the reference & the message shown here indicates when this is so.

Test build added to test adding sequences to an existing alignment.

See https://mafft.cbrc.jp/alignment/software/addsequences.html for more information about the alignment algorithm.