Check branch labels upon v2 dataset JSON validation
If a default branch label is set in the JSON then ensure it is defined on at least one node in the tree. We allow the special case of setting the default label to `none`. Note that `augur export` currently only exports clade and/or aa labels, but we should still be able to validate JSONs made somewhere else.

A test build is added.