Require a version of Snakemake before 5.11 in the Conda environment
Version 5.11 made the --cores option mandatory, which breaks most
existing invocations of snakemake in our workflows (and the workflow
patterns we recommend).  This is causing Travis CI failures, as noted in
issue #456.

The previous commit makes the same version change for setup.py for when
augur is installed with pip.

If Snakemake reverts this breaking change¹, then we can revert this
change as well, while declaring 5.11.0 and 5.11.1 to be incompatible.

¹ https://github.com/snakemake/snakemake/issues/284