MNT: use raw strings for regular expressions
In Python strings '\' is the escape to enter some special
characters (ex '\n' for new line and '\t' for tab).  Previously
if '\' was not followed by a letter that made one of the escape
sequences Python would shrug and assume you really meant just '\',
however this leads to lots of silent errors (when you type \m instead
of \n or tau needs to be written \\tau) so now Python is warning and
in the future will fail if there is an invalid escape sequence.  By
making thees strings 'raw' (the r prefix) it tells Python to never
treat '\' in that string as the start of an escape sequence.