Allow exception to bubble up when priority file is bad.
When the user specifies a priority file, the program should either obey
the user's instructions or fail. A "best effort" approach with default
values may result in the program doing something that the user did not
intend.

With "fail on error" behavior, user can have confidence that if the
program completes, it followed the instructions as specified.