docs: Quote package extras argument in examples
These work unquoted in bash, but not in zsh (and maybe other shells).
Zsh is sufficiently common that we should account for it.

The unquoted .[dev] is treated as glob pattern.  It typically will fail
expand because there aren't .d, .e, or .v files committed, and bash
passes through unexpanded globs as-is by default.  zsh, on the other
hand, defaults to throwing an error on failed globs.  Quoting means it
isn't syntactically a glob anymore.