First attempts to add support for mafft's reference-based alignment
Adds a `--add-fragments` flag to enable mafft's `--addfragments` mode that turns
on a reference-based alignment instead of a full multiple sequence
alignment.

This is a work in progress. One issue is that mafft requires the reference input
to be a FASTA file (it crashes on a GenBank input), but we have some lingering
issues with augur align and how FASTA vs. GenBank references are handled. This
means that the `--remove-reference` flag doesn't work as expected for initial
testing. However, using the new flag from this commit does produce an alignment
using the new mode as expected.

We also need tests for this new mode.