Add cram to test runner script
Run functional tests when running the generic "run_tests" script and when full
unit tests are also being run. We skip functional tests when the user requests a
subset of unit tests to facilitate rapid testing during test-driven development.

Cram tests use pushd and popd which are bash-specific and not available in
Travis's default shell, so we specify the shell for Cram tests to use.