print treetime version
Depending on the install mechanism it seems possible to have an older-than-asked-for treetime
version in use, which can lead to confusing errors. Here we print the version of treetime
used for the commands which use it (excluding those which use a helper command such as `read_vcf`)
This is printed before the treetime command is run so that it's visible if that fails.