Remove snakemake as a dependency
Technically augur does not depend on Snakemake to run and including it here can
cause confusion and/or issues installing either augur or Snakemake. This commit
removes the snakemake dependency and leaves the installation of Snakemake up to
the user.

Note that functional tests continue to rely on Snakemake, so we need to install
it as a dev dependency. Once functional tests have migrated completely to cram,
we can remove this dependency. We use the latest minor version of the 5.*
Snakemake releases, now that our documentation has been updated to reflect the
required `--cores` argument.

This commit also removes the Snakemake dependency from the conda environment
file, since the version installed from bioconda is ancient (circa 3.*) and gets
replaced by the pip installed version from our dev environment anyway.