Update schemas to allow display defaults for transmission lines
Schemas and test builds updated to allow one to define a display default
in a dataset JSON to toggle on/off the transmission lines (if available).
See the corresponding auspice PR https://github.com/nextstrain/auspice/pull/1152.

Additionally, the logic for setting display_defaults within `augur export v2` is
improved so that we no longer have to define the valid keys in the source code.
Invalid keys are still identified via validating the JSONs against schemas.