Forbid constrained digits less significant than uncertain digits.
augur.utils.ambiguous_date_to_date_range resolves ambiguous dates into
tuples of the min and max possible dates. Uncertainty is conveyed by
using the "X" character in the date pattern.

But it usually only makes sense for uncertainty to be in the
least-significant places. For example, assuming %Y-%m-%d, a pattern of
2000-XX-05 would mean the 5th day of any month in 2000. This is almost
certainly erroneous data and undesired behavior.

Therefore, DateDisambiguator now raises an error when a constrained
digit appears in a less-significant place than an uncertain digit.