Match bounds for observations in frequency estimator classes
Resolves an edge case bug where there are enough observations to pass the
upstream filters of the `freq_est_clipped` class but not enough observations to
pass the `frequency_estimator` class's more restrictive lower bound. This bug
manifests as an attempt to interpolate with a single observation.

This commit resolves the issue by changing the lower bound test of the
`frequency_estimator` class to match the upstream lower bound of the
`freq_est_clipped` class.