Prototype interface for grouping frequencies by metadata field
Adds an optional `--group-by` argument to augur frequencies and implements
initial logic to group strains by the given metadata column name and calculate
diffusion-based mutation frequencies per group.

This is a very rough prototype. As such it does not work for the KDE method of
mutation frequencies or any of the tree frequencies and needs tests.