Sync and deduplicate installation instructions
Replaces installation instructions from the README with a link to the full
installations instructions in the docs. Also, adds a link out to the Zika
tutorial to flesh out the "quickstart". Updates the full installation
instructions with a missing brew tap command that was adding to the README in an
earlier commit.

Resolves #519