Flexibly disambiguate multiple publications by the same author
Adds a function to convert a zero-indexed article count for a given author to a
alphabetical disambiguation suffix. For example, two articles by Bedford et
al. get disambiguated as "Bedford et al. A" and "Bedford et al. B".

The new function is slightly overkill in that it supports an infinite number of
articles per author. We could easily unroll the first two loops into redundant
code that only works for a fixed number of articles (702), but the idea here is
that we never have to touch this function again.

Fixes #571