Allow sample-per-group to be dynamically inferred given a maximum number of samples
Due to growing datasets, we want to be able to dynamically adjust the sampling rate.  If we provide a fixed sampling rate, our nextstrain runs grow longer and longer.  We want to specify a target number of samples and pick the spg that will produce the closest number of samples that is still less than the target.