pass -nt AUTO -ntmax {nthreads} for IQ-Tree
specify "-nt AUTO" to tell IQ-TREE to automatically determine the best number of cores given the current data, with an upper bound of "-ntmax nthreads", the number of cores specified by the user. This puts the onus on IQ-TREE to determine the number of threads to use within the limit specified by the user. The actual number of threads usedby IQ-TREE may be lower than the value of nthreads, depending on the data. This avoid a scenario where IQ-TREE crashes because the user has specified more cores than IQ-TREE can use for a given dataset. See:
https://github.com/nextstrain/augur/issues/589
https://github.com/nextstrain/augur/pull/590
This also applies `shquote` to the `log_file` path in the event a substitution model is specified (previously it was only applied if no model was given).