Update functional tests to expect deprecation warning
The latest version of numpy issues a deprecation warning when it encounters
ragged nested arrays. One such array gets imported implicitly from TreeTime
when running augur refine and ancestral. Until a fix for this is merged in
TreeTime, we change our expectations for augur output to reflect this warning
and prevent CI builds from breaking over a warning.