Update call and expected returns from pandas's parse_time_string
The API for `parse_time_string` changed recently to only return two values
instead of three [1] and the function is no longer importable from the pandas
core tools datetimes module as of version 1.1.0. Instead, we have to import the
`parsing` module that exposes the `parse_time_string` function. This commit
supports both the previous and new API's return values.

This bug should affect anyone trying to run augur with pandas 1.1.0 but not
users running earlier versions of pandas.

[1] https://github.com/pandas-dev/pandas/pull/31065