Add minimal unit tests for titer models
Adds unit tests with a reasonable subset of previously published titer data,
amino acid sequence alignments, and a tree. These tests allow us to remove the
`__main__` code from the `titer_model` module that was a holdover from an
ancient version of augur and would not run anyway.

As a side effect of testing the `validate` methods of the two titer models, we
also identified a Python 2to3 bug where dictionary values were not being
properly converted to lists. This commit also fixes that bug.