Add functional tests for augur refine
Adds functional tests based on the Zika build including commands that
run the usual time tree inference, use refine to label internal nodes
without running time tree, and use refine to specify the type of
divergence unit to report (mutations or mutations per site).

Note that the last of these tests currently fails because the command
asks for mutations as the divergence unit, but the input does not
include an alignment FASTA. In this case, augur refine should throw an
error alerting the user that they are missing a required input.