[fetch-docs] relative links
this relative link did not work
after fetching this doc into
docs.nextstrain.org because the
folder structure is different from
this repo so this changes it to fit
with that of docs.nextstrain.org.
this means the relative link breaks
in this repo but this document is
not rendered in this repo's docs.
This will need to remain this way
or become an absolute link until
we have a unified toctree.