Build workflow figure with dot language
Replaces the hardcoded PDF figure for the manuscript with a figure
generated by a Makefile using the dot executable from the Graphviz
library and a single dot file with both Zika and TB workflows
represented as subgraphs. This approach allows us to easily modify
aspects of the figure by modifying the dot source instead of modifying
static PDFs and automatically annotate subfigure labels (A and B).

Adding a Makefile also makes recompiling the paper easier. Along with
the Makefile, this commit updates the README to include instructions for
making a conda environment with the dependencies required to compile the
manucript.